+++ axum/src/body/mod.rs
@@ -1,10 +1,10 @@
 //! HTTP body utilities.
 
 #[doc(no_inline)]
-pub use http_body::{Body as HttpBody, Empty, Full};
+pub use http_body::Body as HttpBody;
 
 #[doc(no_inline)]
 pub use bytes::Bytes;
 
 #[doc(inline)]
-pub use axum_core::body::{boxed, Body, BoxBody};
+pub use axum_core::body::Body;