+++ axum/src/boxed.rs
@@ -1,7 +1,6 @@
 use std::{convert::Infallible, fmt};
 
-use axum_core::body::Body;
-use http::Request;
+use crate::extract::Request;
 use tower::Service;
 
 use crate::{
@@ -37,11 +36,7 @@ where
         }))
     }
 
-    pub(crate) fn call_with_state(
-        self,
-        request: Request<Body>,
-        state: S,
-    ) -> RouteFuture<Infallible> {
+    pub(crate) fn call_with_state(self, request: Request, state: S) -> RouteFuture<Infallible> {
         self.0.call_with_state(request, state)
     }
 }
@@ -82,7 +77,7 @@ pub(crate) trait ErasedIntoRoute<S, E>: Send {
 
     fn into_route(self: Box<Self>, state: S) -> Route<E>;
 
-    fn call_with_state(self: Box<Self>, request: Request<Body>, state: S) -> RouteFuture<E>;
+    fn call_with_state(self: Box<Self>, request: Request, state: S) -> RouteFuture<E>;
 }
 
 pub(crate) struct MakeErasedHandler<H, S> {
@@ -103,11 +98,7 @@ where
         (self.into_route)(self.handler, state)
     }
 
-    fn call_with_state(
-        self: Box<Self>,
-        request: Request<Body>,
-        state: S,
-    ) -> RouteFuture<Infallible> {
+    fn call_with_state(self: Box<Self>, request: Request, state: S) -> RouteFuture<Infallible> {
         self.into_route(state).call(request)
     }
 }
@@ -141,11 +132,7 @@ where
         (self.into_route)(self.router, state)
     }
 
-    fn call_with_state(
-        mut self: Box<Self>,
-        request: Request<Body>,
-        state: S,
-    ) -> RouteFuture<Infallible> {
+    fn call_with_state(mut self: Box<Self>, request: Request, state: S) -> RouteFuture<Infallible> {
         self.router.call_with_state(request, state)
     }
 }
@@ -184,7 +171,7 @@ where
         (self.layer)(self.inner.into_route(state))
     }
 
-    fn call_with_state(self: Box<Self>, request: Request<Body>, state: S) -> RouteFuture<E2> {
+    fn call_with_state(self: Box<Self>, request: Request, state: S) -> RouteFuture<E2> {
         (self.layer)(self.inner.into_route(state)).call(request)
     }
 }